requirements_hf.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
yfinance
pandas
numpy
scikit-learn
requests
beautifulsoup4
transformers
torch
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt==4.0.1
psycopg2-binary
httpx